requirements.txt
joblib
numpy
pillow
streamlit
dill
tensorflow
keras